+++ CHANGELOG.md
@@ -11,7 +11,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Changed
 * Experimental feature where the CLI can now accept `_` or `-` in sys.argv if `__fuzzy_hyphens__` exists and is truthy
-
+* default, description, and epilog can now be specified with dunder __default__, __description__, and __epilog__ attributes
+ 
 ### Fixed
 * Bug where trying to get a non-existing value raised an AttributeError instead of a KeyError due to aliases
 * Fixed issue where using setattr on a DataConfig using a known key did not set its dictionary value.